+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to Merlin Spellbook will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+## Added
+- Delimiter option in serialization command.
+
+## Changed
+- In serialize command, grab the first entry as the key and everything else as the value.
+
 ## [0.5.3]
 
 ## Fixed